requirements_professional.txt
# MCP Web Scraper Professional Requirements
# Core MCP and async support
mcp>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Web scraping and browser automation  
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0

# Image processing and computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
face-recognition>=1.3.0
numpy>=1.24.0

# Machine learning and data processing
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0

# Data handling and storage
pandas>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities and validation
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorlog>=6.7.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
